+++ src/main/java/io/github/petty/llm/service/EmbeddingService.java
@@ -87,6 +87,8 @@ public class EmbeddingService {
         metadata.put("areaCode", content.getAreaCode());
         // sigunguCode (시군구 추가)
         metadata.put("sigunguCode", content.getSigunguCode());
+        // 지역 String 추가
+        metadata.put("address", content.getAddr1());
         // contentType (콘텐츠 관련 추가)
         metadata.put("contentType", content.getContentTypeId());
