+++ src/main/java/io/github/petty/llm/service/RecommendService.java
@@ -1,0 +1,7 @@
+package io.github.petty.llm.service;
+
+import java.util.Map;
+
+public interface RecommendService {
+    String recommend(Map<String, String> promptMap);
+}
